[Edit Hotel Info] Add flow edit hotel info

+++ app/src/main/java/com/minhhnn18898/letstravel/tripdetail/data/model/FlightInfoModel.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class FlightInfoModel(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo("flight_id")
-    val flightId: Int,
+    val flightId: Long,
     @ColumnInfo("flight_number")
     val flightNumber: String,
     @ColumnInfo("operated_airlines")
